requirements.txt
numpy>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
torch>=2.0.0
torchaudio>=2.0.0
gradio>=4.0.0
chatterbox-tts
